patches and bug fixes back to the version 1 code, before the GUI and the namespace changes

+++ include/Calibration.hh
@@ -38,12 +38,12 @@ public:
 		return fInputFile;
 	}
 	float AsicEnergy( unsigned int mod, unsigned int asic, unsigned int chan, unsigned short raw );
-	float AsicThreshold( unsigned int mod, unsigned int asic, unsigned int chan );
+	unsigned int AsicThreshold( unsigned int mod, unsigned int asic, unsigned int chan );
 	long AsicTime( unsigned int mod, unsigned int asic );
 	bool AsicEnabled( unsigned int mod, unsigned int asic );
 	float AsicWalk( unsigned int mod, unsigned int asic, float energy );
 	float CaenEnergy( unsigned int mod, unsigned int chan, unsigned short raw );
-	float CaenThreshold( unsigned int mod, unsigned int chan );
+    unsigned int CaenThreshold( unsigned int mod, unsigned int chan );
 	long CaenTime( unsigned int mod, unsigned int chan );
 	
 
@@ -57,7 +57,7 @@ private:
 	std::vector< std::vector< std::vector<float> > > fAsicOffset;
 	std::vector< std::vector< std::vector<float> > > fAsicGain;
 	std::vector< std::vector< std::vector<float> > > fAsicGainQuadr;
-	std::vector< std::vector< std::vector<float> > > fAsicThreshold;
+	std::vector< std::vector< std::vector<unsigned int> > > fAsicThreshold;
 	std::vector< std::vector<long> > fAsicTime;
 	std::vector< std::vector<bool> > fAsicEnabled;
 	std::vector< std::vector< std::vector<float> > > fAsicWalk;
@@ -65,7 +65,7 @@ private:
 	std::vector< std::vector<float> > fCaenOffset;
 	std::vector< std::vector<float> > fCaenGain;
 	std::vector< std::vector<float> > fCaenGainQuadr;
-	std::vector< std::vector<float> > fCaenThreshold;
+	std::vector< std::vector<unsigned int> > fCaenThreshold;
 	std::vector< std::vector<long> > fCaenTime;
 
 	//ClassDef(Calibration, 1)
